requirements.txt
pandas
openpyxl
chromadb
sentence-transformers
google-genai
gradio
pdfplumber
python-dotenv
streamlit
python-dotenv
chromadb
